Add CMake installation rules.
This makes it so you can type `make install` and install the library and
Python module in the usual way.

We do *not* install the C headers, so it's not actually possible to write C
code that links against libsymphony. That should be fixed, and shouldn't be
hard to fix.

I made up a version number of 0.1 for the shared library. Note that there are
specific semantics to shared library versions that must be respected when
making changes. On the other hand, the shared library version number need not
be identical to the overall package version number.