src/power_law/power_law.c: actually zero out the distribution function outside of gamma_{min,max}
This doesn't come up in most circumstances, but it is relevant if you try to
calculate coefficients at or below the cyclotron frequency.

For now, `differential_of_power_law` returns NAN if you try to go beyond the
bounds of the gamma cutoffs, since the discontinuity in the function makes the
derivative become undefined, and any integration extending past the cutoffs
should have to care about that. Note that the analytic differential function
is not actually hooked up to any Python code at the moment, though.